requirements.txt
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31
cos-python-sdk-v5>=1.9
